fix: auto generate declaration file
