+++ package.json
@@ -9,7 +9,7 @@
   "author": "Justin Schrader <me@justin.beer>",
   "license": "MIT",
   "private": false,
-  "types": "types/use-react-router-breadcrumbs/index.d.ts",
+  "types": "dist/index.d.ts",
   "peerDependencies": {
     "react": ">=16.8",
     "react-router": ">=5.1.0"
@@ -19,9 +19,7 @@
     "build": "rollup -c",
     "test": "jest",
     "test-build": "sh ./scripts/test-build.sh",
-    "types": "yarn type-src && yarn type-descriptions",
-    "type-src": "tsc -p tsconfig.json",
-    "type-descriptions": "tsc -p types/use-react-router-breadcrumbs",
+    "types": "tsc -p tsconfig.json",
     "lint": "eslint ./src/**"
   },
   "husky": {
@@ -66,6 +64,7 @@
     "react-router": "^5.1.2",
     "rollup": "^2.1.0",
     "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-typescript2": "^0.27.0",
     "rollup-plugin-uglify": "^6.0.2",
     "typescript": "^3.8.3"
   },
