+++ tsconfig.json
@@ -4,7 +4,10 @@
     "checkJs": true,
     "esModuleInterop": true,
     "jsx": "react",
-    "lib": ["es2015", "dom"],
+    "lib": [
+      "es2015",
+      "dom"
+    ],
     "module": "commonjs",
     "moduleResolution": "node",
     "noEmit": true,
@@ -17,7 +20,9 @@
     "strictFunctionTypes": true,
     "strictNullChecks": true,
     "strictPropertyInitialization": true,
-    "target": "es5"
+    "target": "es5",
+    "declaration": true,
+    "declarationDir": "./dist"
   },
   "files": [
     "src/index.tsx"
