--- types/use-react-router-breadcrumbs/index-tests.tsx
@@ -1,10 +1,0 @@
-/* eslint-disable import/no-unresolved */
-/* eslint-disable import/extensions */
-import * as React from 'react';
-import BreadCrumbs from './use-react-router-breadcrumbs-tests';
-
-export default function () {
-  return (
-    <BreadCrumbs />
-  );
-}
