--- types/use-react-router-breadcrumbs/tsconfig.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "../",
-    "forceConsistentCasingInFileNames": true,
-    "jsx": "react",
-    "lib": [
-      "es6",
-      "dom"
-    ],
-    "module": "commonjs",
-    "noEmit": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "strictNullChecks": true,
-    "typeRoots": [
-      "../"
-    ],
-    "types": []
-  },
-  "files": [
-    "index.d.ts",
-    "index-tests.tsx",
-    "use-react-router-breadcrumbs-tests.tsx"
-  ]
-}
